# CUDA: Faster FlashAttention kernel