This PR adds a rewritten, faster FlashAttention CUDA kernel. A notable feature is that tensor core fragments can now have different shapes depending on batch size. Performance for batch sizes 1-4 is still somewhat suboptimal, especially for head sizes 64 and 80. I think there is no way around writing a second FlashAttention kernel that does not use tensor cores (luckily this will be comparatively simple though). Performance on my system changes as follows:

|   Batch size |   t/s ~~master~~ jg /flash-attn-12 no FlashAttention |   t/s gg/flash-attn |   Speedup gg/flash-attn |   t/s jg/flash-attn-12 |   Speedup jg/flash-attn-12 |
|-------------:|-------------:|--------------------:|------------------------:|-----------------------:|---------------------------:|
|         1 |       112.51 |              104.20 |                    0.93 |                 110.48 |                       0.98 |
|         2 |       212.14 |              203.85 |                    0.96 |                 215.48 |                       1.02 |
|         4 |       338.06 |              349.16 |                    1.03 |                 365.31 |                       1.08 |
|         8 |       444.67 |              488.77 |                    1.10 |                 504.61 |                       1.13 |
|        16 |       450.71 |              519.93 |                    1.15 |                 529.54 |                       1.17 |
|        32 |       578.73 |              536.66 |                    0.93 |                 582.85 |                       1.01 |
|        64 |      1183.09 |             1156.06 |                    0.98 |                1273.54 |                       1.08 |
|       128 |      1891.63 |             1979.86 |                    1.05 |                2120.92 |                       1.12 |
|       256 |      2564.85 |             2861.33 |                    1.12 |                2943.26 |                       1.15 |
|       512 |      2934.15 |             3279.97 |                    1.12 |                3433.73 |                       1.17 |
|      1024 |      2938.55 |             3293.07 |                    1.12 |                3445.01 |                       1.17 |
|      2048 |      2945.06 |             3311.10 |                    1.12 |                3457.93 |                       1.17 |
|      4096 |      2951.00 |             3321.18 |                    1.13 |                3468.74 |                       1.18 |

Plot of the same numbers:

![flash_attention_perf](https://github.com/ggerganov/llama.cpp/assets/18492268/e65c0c5c-560a-49d8-833d-f13969feabfe)

Initially I tried an implementation that calculated softmax using the maximum value of multiple KQ columns. This allowed you to directly scale tensor core fragments instead of having to go through shared memory. This was faster than the kernel in this PR but unfortunately also had a ~0.1% chance to blow up the computation and give you NaN.

For my kernel I am assuming `K->ne[1] % D == 0`. Unless I am missing something something this should always be true for actual use in language models. However, the tests included cases where this is not the case. For this reason I have edited the tests.