@JohannesGaessler 2024-03-28
>Yes, this can be satisfied by padding the KV cache appropriately

No, I don't mean the size of the KV cache. I mean the size of the view of the KV cache used in the kernel. Unless I'm missing something that view is always a multiple of n_embd_head_k == D and I am explicitly making use of that fact.