@JohannesGaessler 2024-03-29
Gemma should produce correct results now. In addition to the padding I had forgotten to consider GQA.