@JohannesGaessler 2024-03-29
Which models have n_embd_head == 64 or 80?