@phymbert 2024-03-29
I would like to see the new `Benchmark` job on this code. Dont you ?