@ggerganov 2024-03-29
The `phi-2` model has head size of 80. However, the F16 precision for this model is not enough and it will produce garbage. We need FA kernels with F32 accumulators that will be used when `GGML_PREC_F32` is specified:

https://github.com/ggerganov/llama.cpp/blob/bfe7dafc9cf96b9a09ead347fed9a547930fc631/llama.cpp#L5705-L5709