@JohannesGaessler 2024-03-29
I am getting correct results for `tests/test-backend-ops` with a head size of 80. However, the actual Phi-2 model is still producing garbage. I had a similar issue with Gemma because that model was using GQA. Does Phi-2 do anything special with its attention mechanism?