@JohannesGaessler 2024-03-29
I'm dumb, I forgot about the FP16 accumulator issues described just one post prior.