@ggerganov 2024-03-29
I'm comparing to the `master` branch. Here are the results with the following command:

```bash
LLAMA_CUBLAS=1 make -j llama-bench && ./llama-bench -m /mnt/llama.cpp/models/gemma-2b/ggml-model-f16.gguf -ngl 99 -p 512,1024,2048
```

ggml_cuda_init: GGML_CUDA_FORCE_MMQ:   no
ggml_cuda_init: CUDA_USE_TENSOR_CORES: yes
ggml_cuda_init: found 1 CUDA devices:
  Device 0: Tesla V100-PCIE-16GB, compute capability 7.0, VMM: yes
| model                          |       size |     params | backend    | ngl | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------- | ---------------: |
| gemma 2B F16 (guessed)         |   5.64 GiB |     3.03 B | CUDA       |  99 | pp 512     | 11308.07 ± 144.58 |
| gemma 2B F16 (guessed)         |   5.64 GiB |     3.03 B | CUDA       |  99 | pp 1024    | 11621.09 ± 82.05 |
| gemma 2B F16 (guessed)         |   5.64 GiB |     3.03 B | CUDA       |  99 | pp 2048    | 11324.10 ± 109.07 |
| gemma 2B F16 (guessed)         |   5.64 GiB |     3.03 B | CUDA       |  99 | tg 128     |    118.77 ± 0.31 |

build: cfde806e (2575)


ggml_cuda_init: GGML_CUDA_FORCE_MMQ:   no
ggml_cuda_init: CUDA_USE_TENSOR_CORES: yes
ggml_cuda_init: found 1 CUDA devices:
  Device 0: Tesla V100-PCIE-16GB, compute capability 7.0, VMM: yes
| model                          |       size |     params | backend    | ngl | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------- | ---------------: |
| gemma 2B F16 (guessed)         |   5.64 GiB |     3.03 B | CUDA       |  99 | pp 512     |  9160.73 ± 69.51 |
| gemma 2B F16 (guessed)         |   5.64 GiB |     3.03 B | CUDA       |  99 | pp 1024    |  8885.20 ± 18.08 |
| gemma 2B F16 (guessed)         |   5.64 GiB |     3.03 B | CUDA       |  99 | pp 2048    |  8117.28 ± 20.89 |
| gemma 2B F16 (guessed)         |   5.64 GiB |     3.03 B | CUDA       |  99 | tg 128     |    122.19 ± 1.71 |

build: e4badc16 (2576)