@JohannesGaessler 2024-03-29
Relative to master I'm also seeing a performance regression for Gemma. But I am seeing this performance regression both for CPU and CUDA and regardless of whether or not `LLAMA_FLASH_ATTENTION` is defined in llama.cpp; If `LLAMA_FLASH_ATTENTION` is defined the performance is comparatively better. To me this suggests that the cause for the regression is some other change either on master or in `gg/flash-attn`; I don't have a comprehensive overview of all the things that are being changed in `gg/flash-attn` and which of those changes are supposed to be temporary.