@ggerganov 2024-03-29
I see - this branch does not have the recent optimization for skipping the computation of the unused tokens in the last layer. Likely after rebasing to latest `gg/flash-attn` the discrepancy will disappear