@JohannesGaessler 2024-03-29
I added a specialized kernel for batch size 1 and rebased on top of the most recent `gg/flash-attn`. This is the current performance:

| GPU      | Model                      |   Model Size [GiB] |   Num. of Parameters |   Batch size | Test   |   t/s master |   t/s jg/flash-attn-16 |   Speedup |
|:---------|:---------------------------|-------------------:|---------------------:|-------------:|:-------|-------------:|-----------------------:|----------:|
| RTX 3090 | gemma 2B all F32 (guessed) |              11.29 |           3030984704 |            1 | pp4096 |        73.10 |                  73.00 |      1.00 |
| RTX 3090 | gemma 2B all F32 (guessed) |              11.29 |           3030984704 |            2 | pp4096 |       128.68 |                 121.51 |      0.94 |
| RTX 3090 | gemma 2B all F32 (guessed) |              11.29 |           3030984704 |            4 | pp4096 |       252.66 |                 240.55 |      0.95 |
| RTX 3090 | gemma 2B all F32 (guessed) |              11.29 |           3030984704 |            8 | pp4096 |       492.13 |                 476.25 |      0.97 |
| RTX 3090 | gemma 2B all F32 (guessed) |              11.29 |           3030984704 |           16 | pp4096 |       940.09 |                 943.43 |      1.00 |
| RTX 3090 | gemma 2B all F32 (guessed) |              11.29 |           3030984704 |           32 | pp4096 |      1902.00 |                1841.48 |      0.97 |
| RTX 3090 | gemma 2B all F32 (guessed) |              11.29 |           3030984704 |           64 | pp4096 |      3408.40 |                3349.84 |      0.98 |
| RTX 3090 | gemma 2B all F32 (guessed) |              11.29 |           3030984704 |          128 | pp4096 |      3002.53 |                3036.20 |      1.01 |
| RTX 3090 | gemma 2B all F32 (guessed) |              11.29 |           3030984704 |          256 | pp4096 |      4159.57 |                4343.17 |      1.04 |
| RTX 3090 | gemma 2B all F32 (guessed) |              11.29 |           3030984704 |          512 | pp4096 |      5746.50 |                6003.39 |      1.04 |
| RTX 3090 | gemma 2B all F32 (guessed) |              11.29 |           3030984704 |         1024 | pp4096 |      5821.18 |                6072.62 |      1.04 |
| RTX 3090 | gemma 2B all F32 (guessed) |              11.29 |           3030984704 |         2048 | pp4096 |      5888.14 |                6136.36 |      1.04 |
| RTX 3090 | gemma 2B all F32 (guessed) |              11.29 |           3030984704 |         4096 | pp4096 |      5912.42 |                6161.33 |      1.04 |
| RTX 3090 | llama 7B Q4_0              |               3.56 |           6738415616 |            1 | pp4096 |       112.50 |                 117.42 |      1.04 |
| RTX 3090 | llama 7B Q4_0              |               3.56 |           6738415616 |            2 | pp4096 |       212.92 |                 213.85 |      1.00 |
| RTX 3090 | llama 7B Q4_0              |               3.56 |           6738415616 |            4 | pp4096 |       342.69 |                 366.36 |      1.07 |
| RTX 3090 | llama 7B Q4_0              |               3.56 |           6738415616 |            8 | pp4096 |       456.29 |                 514.65 |      1.13 |
| RTX 3090 | llama 7B Q4_0              |               3.56 |           6738415616 |           16 | pp4096 |       466.84 |                 553.58 |      1.19 |
| RTX 3090 | llama 7B Q4_0              |               3.56 |           6738415616 |           32 | pp4096 |       606.66 |                 625.60 |      1.03 |
| RTX 3090 | llama 7B Q4_0              |               3.56 |           6738415616 |           64 | pp4096 |      1212.55 |                1311.68 |      1.08 |
| RTX 3090 | llama 7B Q4_0              |               3.56 |           6738415616 |          128 | pp4096 |      1929.57 |                2181.56 |      1.13 |
| RTX 3090 | llama 7B Q4_0              |               3.56 |           6738415616 |          256 | pp4096 |      2606.33 |                3040.31 |      1.17 |
| RTX 3090 | llama 7B Q4_0              |               3.56 |           6738415616 |          512 | pp4096 |      2975.79 |                3542.18 |      1.19 |
| RTX 3090 | llama 7B Q4_0              |               3.56 |           6738415616 |         1024 | pp4096 |      2981.22 |                3538.68 |      1.19 |
| RTX 3090 | llama 7B Q4_0              |               3.56 |           6738415616 |         2048 | pp4096 |      2983.94 |                3549.31 |      1.19 |
| RTX 3090 | llama 7B Q4_0              |               3.56 |           6738415616 |         4096 | pp4096 |      2986.36 |                3547.84 |      1.19 |
| RTX 3090 | phi2 3B F16                |               5.18 |           2779683840 |            1 | pp4096 |       100.57 |                  71.66 |      0.71 |
| RTX 3090 | phi2 3B F16                |               5.18 |           2779683840 |            2 | pp4096 |       175.81 |                 130.96 |      0.74 |
| RTX 3090 | phi2 3B F16                |               5.18 |           2779683840 |            4 | pp4096 |       340.64 |                 260.00 |      0.76 |
| RTX 3090 | phi2 3B F16                |               5.18 |           2779683840 |            8 | pp4096 |       656.95 |                 514.21 |      0.78 |
| RTX 3090 | phi2 3B F16                |               5.18 |           2779683840 |           16 | pp4096 |      1206.90 |                1005.32 |      0.83 |
| RTX 3090 | phi2 3B F16                |               5.18 |           2779683840 |           32 | pp4096 |      2130.25 |                1911.41 |      0.90 |
| RTX 3090 | phi2 3B F16                |               5.18 |           2779683840 |           64 | pp4096 |      3448.82 |                3191.17 |      0.93 |
| RTX 3090 | phi2 3B F16                |               5.18 |           2779683840 |          128 | pp4096 |      4756.77 |                4570.06 |      0.96 |
| RTX 3090 | phi2 3B F16                |               5.18 |           2779683840 |          256 | pp4096 |      5279.77 |                4930.82 |      0.93 |
| RTX 3090 | phi2 3B F16                |               5.18 |           2779683840 |          512 | pp4096 |      5665.75 |                5178.48 |      0.91 |
| RTX 3090 | phi2 3B F16                |               5.18 |           2779683840 |         1024 | pp4096 |      5673.42 |                5149.88 |      0.91 |
| RTX 3090 | phi2 3B F16                |               5.18 |           2779683840 |         2048 | pp4096 |      5649.56 |                5140.25 |      0.91 |
| RTX 3090 | phi2 3B F16                |               5.18 |           2779683840 |         4096 | pp4096 |      5669.77 |                5113.27 |      0.90 |

LLaMA (head size 128) is universally faster. Gemma (head size 256) is faster for the most important use cases of batch size 1 and batch size >= 512. Phi-2 (head size 80) is slower because 80 is just a terrible choice of head size for CUDA. Quite honestly I don't want to put in the effort to optimize for that particular head size because as far as I'm concerned 3b models are pretty much useless.