@JohannesGaessler 2024-04-01
I implemented splitting the calculations across multiple CUDA blocks for batch size 1 in order to get better GPU utilization; the same trick is also being used on the official FlashAttention repository. This is the current performance on my system:

<details>

| GPU        | Model           |     Batch size | Test     |     t/s master |     t/s jg/flash-attn-18 |     Speedup |
| :--------- | :-------------- | -------------: | :------- | -------------: | -----------------------: | ----------: |
| RTX 3090   | gemma 2B F16    |              1 | pp4096   |         124.27 |                   135.22 |        1.09 |
| RTX 3090   | gemma 2B F16    |              2 | pp4096   |         223.37 |                   200.10 |        0.90 |
| RTX 3090   | gemma 2B F16    |              4 | pp4096   |         446.92 |                   400.69 |        0.90 |
| RTX 3090   | gemma 2B F16    |              8 | pp4096   |         861.96 |                   792.99 |        0.92 |
| RTX 3090   | gemma 2B F16    |             16 | pp4096   |        1588.75 |                  1553.48 |        0.98 |
| RTX 3090   | gemma 2B F16    |             32 | pp4096   |        3181.24 |                  2920.65 |        0.92 |
| RTX 3090   | gemma 2B F16    |             64 | pp4096   |        5742.32 |                  5285.37 |        0.92 |
| RTX 3090   | gemma 2B F16    |            128 | pp4096   |        9333.71 |                  9239.47 |        0.99 |
| RTX 3090   | gemma 2B F16    |            256 | pp4096   |       10957.27 |                 11859.30 |        1.08 |
| RTX 3090   | gemma 2B F16    |            512 | pp4096   |       11947.09 |                 13296.03 |        1.11 |
| RTX 3090   | gemma 2B F16    |           1024 | pp4096   |       12135.52 |                 13492.08 |        1.11 |
| RTX 3090   | gemma 2B F16    |           2048 | pp4096   |       12252.33 |                 13748.61 |        1.12 |
| RTX 3090   | gemma 2B F16    |           4096 | pp4096   |       12420.43 |                 13626.31 |        1.10 |
| RTX 3090   | llama 7B Q4_0   |              1 | pp4096   |         113.18 |                   125.32 |        1.11 |
| RTX 3090   | llama 7B Q4_0   |              2 | pp4096   |         213.52 |                   224.22 |        1.05 |
| RTX 3090   | llama 7B Q4_0   |              4 | pp4096   |         346.91 |                   382.54 |        1.10 |
| RTX 3090   | llama 7B Q4_0   |              8 | pp4096   |         462.59 |                   534.06 |        1.15 |
| RTX 3090   | llama 7B Q4_0   |             16 | pp4096   |         470.46 |                   567.01 |        1.21 |
| RTX 3090   | llama 7B Q4_0   |             32 | pp4096   |         614.74 |                   643.33 |        1.05 |
| RTX 3090   | llama 7B Q4_0   |             64 | pp4096   |        1219.97 |                  1338.04 |        1.10 |
| RTX 3090   | llama 7B Q4_0   |            128 | pp4096   |        1937.68 |                  2277.09 |        1.18 |
| RTX 3090   | llama 7B Q4_0   |            256 | pp4096   |        2627.53 |                  3153.93 |        1.20 |
| RTX 3090   | llama 7B Q4_0   |            512 | pp4096   |        3001.62 |                  3720.68 |        1.24 |
| RTX 3090   | llama 7B Q4_0   |           1024 | pp4096   |        3002.21 |                  3730.97 |        1.24 |
| RTX 3090   | llama 7B Q4_0   |           2048 | pp4096   |        3005.64 |                  3734.92 |        1.24 |
| RTX 3090   | llama 7B Q4_0   |           4096 | pp4096   |        3006.77 |                  3739.41 |        1.24 |
| RTX 3090   | phi2 3B F16     |              1 | pp4096   |         100.77 |                    96.94 |        0.96 |
| RTX 3090   | phi2 3B F16     |              2 | pp4096   |         176.08 |                   165.16 |        0.94 |
| RTX 3090   | phi2 3B F16     |              4 | pp4096   |         343.50 |                   327.68 |        0.95 |
| RTX 3090   | phi2 3B F16     |              8 | pp4096   |         661.02 |                   645.63 |        0.98 |
| RTX 3090   | phi2 3B F16     |             16 | pp4096   |        1214.18 |                  1257.54 |        1.04 |
| RTX 3090   | phi2 3B F16     |             32 | pp4096   |        2148.22 |                  2250.75 |        1.05 |
| RTX 3090   | phi2 3B F16     |             64 | pp4096   |        3493.84 |                  3624.80 |        1.04 |
| RTX 3090   | phi2 3B F16     |            128 | pp4096   |        4805.64 |                  5270.91 |        1.10 |
| RTX 3090   | phi2 3B F16     |            256 | pp4096   |        5337.40 |                  5766.99 |        1.08 |
| RTX 3090   | phi2 3B F16     |            512 | pp4096   |        5703.90 |                  6465.30 |        1.13 |
| RTX 3090   | phi2 3B F16     |           1024 | pp4096   |        5715.14 |                  6392.56 |        1.12 |
| RTX 3090   | phi2 3B F16     |           2048 | pp4096   |        5737.96 |                  6419.13 |        1.12 |
| RTX 3090   | phi2 3B F16     |           4096 | pp4096   |        5745.58 |                  6387.34 |        1.11 |


</details>

The biggest issue is still batch size 1 + head size 80. Quite frankly I don't know what else to do in order to improve performance. @ggerganov how strict is the goal of getting consistently better performance than master? From this point on the amount of time that it will take me to find improvements for small batch sizes (if there are any at all) will be much higher. Also I will need to attend to other projects so I will not have a lot of time to work on this in the immediate future.