@JohannesGaessler 2024-04-02
I fixed the compiler warnings as well as the cmake build. While going through the code I also discovered that there luckily (?) was a bug that caused excessive shared memory loads for head sizes < 128. With the bug fixed Phi-2 performance is much better:

| GPU        | Model           |     Batch size | Test     |     t/s master |     t/s PR |     Speedup |
| :--------- | :-------------- | -------------: | :------- | -------------: | ---------------: | ----------: |
| RTX 3090   | phi2 3B F16     |              1 | pp4096   |         100.77 |           106.13 |        1.05 |
| RTX 3090   | phi2 3B F16     |              2 | pp4096   |         176.08 |           170.33 |        0.97 |
| RTX 3090   | phi2 3B F16     |              4 | pp4096   |         343.50 |           338.29 |        0.98 |
| RTX 3090   | phi2 3B F16     |              8 | pp4096   |         661.02 |           669.26 |        1.01 |
| RTX 3090   | phi2 3B F16     |             16 | pp4096   |        1214.18 |          1286.26 |        1.06 |
| RTX 3090   | phi2 3B F16     |             32 | pp4096   |        2148.22 |          2424.68 |        1.13 |
| RTX 3090   | phi2 3B F16     |             64 | pp4096   |        3493.84 |          4252.78 |        1.22 |
| RTX 3090   | phi2 3B F16     |            128 | pp4096   |        4805.64 |          6858.33 |        1.43 |
| RTX 3090   | phi2 3B F16     |            256 | pp4096   |        5337.40 |          7708.88 |        1.44 |
| RTX 3090   | phi2 3B F16     |            512 | pp4096   |        5703.90 |          8682.26 |        1.52 |
| RTX 3090   | phi2 3B F16     |           1024 | pp4096   |        5715.14 |          8488.52 |        1.49 |
| RTX 3090   | phi2 3B F16     |           2048 | pp4096   |        5737.96 |          8682.73 |        1.51 |
| RTX 3090   | phi2 3B F16     |           4096 | pp4096   |        5745.58 |          8433.07 |        1.47 |

From my end, this PR is now ready to merge. I'll leave merging to you since I don't know whether you want to squash prior to merging. Feel free to go through the code and request additional explanatory comments if necessary.